💄 Style(canvas): sidebar UI 재구성 및 SelectPlatform 초기 ui 구현
